requirements.txt
joblib
numpy
scikit-learn
pandas
streamlit
plotly.express
